refactor(Client): create custom hook for genderField
